Chaining checks of pairs of pointers by AND operations

+++ tests/bad-alias.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <algorithm>
 
-void foo (int *A, int *B, float *C, int n) {
+void foo (int *A, int *B, int *C, int n) {
   for (int i = 0; i < n; i++) {
     A[i] = i;
     B[i] = A[i+2];
@@ -11,7 +11,7 @@ void foo (int *A, int *B, float *C, int n) {
 
 int main () {
   int A[100], B[100];
-  float C[200];
+  int C[200];
 
   foo (A, B, C, 100);
